raw instances can now be launched in xenapi (only as hvm at the moment)

+++ nova/virt/libvirt_conn.py
@@ -448,20 +448,31 @@ class LibvirtConnection(object):
                            'kernel_id': inst['kernel_id'],
                            'ramdisk_id': inst['ramdisk_id']}
         if not os.path.exists(basepath('disk')):
-            yield images.fetch(inst.image_id, basepath('disk-raw'), user,
-                               project)
-        if not os.path.exists(basepath('kernel')):
-            yield images.fetch(inst.kernel_id, basepath('kernel'), user,
-                               project)
-        if not os.path.exists(basepath('ramdisk')):
-            yield images.fetch(inst.ramdisk_id, basepath('ramdisk'), user,
-                               project)
+           yield images.fetch(inst.image_id, basepath('disk-raw'), user, 
+                              project)
+
+        using_kernel = inst.kernel_id
+        if using_kernel:
+            if not os.path.exists(basepath('kernel')):
+                yield images.fetch(inst.kernel_id, basepath('kernel'), user, 
+                                   project)
+            if not os.path.exists(basepath('ramdisk')):
+                yield images.fetch(inst.ramdisk_id, basepath('ramdisk'), user, 
+                                   project)
+        
 
         execute = lambda cmd, process_input = None, check_exit_code = True: \
                   process.simple_execute(cmd=cmd,
                                          process_input=process_input,
                                          check_exit_code=check_exit_code)
 
+        # For now, we assume that if we're not using a kernel, we're using a 
+        # partitioned disk image where the target partition is the first 
+        # partition
+        target_partition = None
+        if not using_kernel:
+            target_partition = "1"
+            
         key = str(inst['key_data'])
         net = None
         network_ref = db.network_get_by_instance(context.get_admin_context(),
@@ -482,12 +493,19 @@ class LibvirtConnection(object):
             if net:
                 logging.info('instance %s: injecting net into image %s',
                     inst['name'], inst.image_id)
-            yield disk.inject_data(basepath('disk-raw'), key, net,
-                                   execute=execute)
+            try:
+                yield disk.inject_data(basepath('disk-raw'), key, net,
+                                       partition=target_partition, 
+                                       execute=execute)
+            except Exception as e:
+                # This could be a windows image, or a vmdk format disk
+                logging.warn('instance %s: ignoring error injecting data'
+                             ' into image %s (%s)',
+                             inst['name'], inst.image_id, e)
 
-        if os.path.exists(basepath('disk')):
-            yield process.simple_execute('rm -f %s' % basepath('disk'))
-
+        if using_kernel:
+            if os.path.exists(basepath('disk')):
+                yield process.simple_execute('rm -f %s' % basepath('disk'))
         local_bytes = (instance_types.INSTANCE_TYPES[inst.instance_type]
                                                     ['local_gb']
                                                     * 1024 * 1024 * 1024)
@@ -524,6 +542,16 @@ class LibvirtConnection(object):
                     'mac_address': instance['mac_address'],
                     'ip_address': ip_address,
                     'dhcp_server': dhcp_server}
+        if instance['kernel_id']:
+            xml_info['kernel'] = xml_info['basepath'] + "/kernel"
+        if instance['ramdisk_id']:
+            xml_info['ramdisk'] = xml_info['basepath'] + "/ramdisk"
+
+        if instance['ramdisk_id'] or instance['kernel_id']:
+            xml_info['disk'] = xml_info['basepath'] + "/disk"
+        else:
+            xml_info['disk'] = xml_info['basepath'] + "/disk-raw"
+        #TODO: Must find a way for not populating kernel and ramdisk attributes
         if rescue:
             libvirt_xml = self.rescue_xml % xml_info
         else:
